requirements.txt
opencv-python
ultralytics
Pillow
SQLAlchemy
